Small refactor, code tidy up
